databricksinternaldemo/mcpserver/requirements.txt
streamlit>=1.28.0
gradio>=4.0.0
databricks-sdk>=0.76.0
databricks-connect>=15.4,<15.5
python-dotenv>=1.0.0
databricks-vectorsearch>=0.1.0
